+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## ScottPlot 5.0.44
 _Not yet on NuGet..._
 * Rendering: Prevent inversion of data area for extremely small plots (#3607)
+* Generate: Created `AddSin()` and `AddSinInPlace()` methods to facilitate test data generation (#3682)
 
 ## ScottPlot 5.0.43
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-11-03_